+++ unittests/scans/veracode/veracode_sca_fixed.json
@@ -1,0 +1,64 @@
+{
+  "_embedded": {
+    "issues": [
+      {
+        "id": "8ef6657a-bd7b-4f18-8107-e5c7b320f718",
+        "site_id": 127637430,
+        "created_date": "2022-07-18T15:49:10.704+0000",
+        "issue_status": "fixed",
+        "issue_type": "vulnerability",
+        "ignored": false,
+        "severity": 6.5,
+        "workspace_id": "cf0a843f-b732-4a2b-ade8-a36ff3ae0181",
+        "project_id": "d5535c41-439a-4d06-a600-4b6da095edda",
+        "project_name": "some-project",
+        "project_branch": "master",
+        "library": {
+          "id": "maven:com.amazonaws:aws-java-sdk-s3:1.11.951:",
+          "name": "AWS Java SDK for Amazon S3",
+          "version": "1.11.951",
+          "release_date": "2021-02-08",
+          "latest_version": "1.12.307",
+          "latest_version_release_date": "2022-09-20",
+          "direct": false,
+          "transitive": true,
+          "_links": {
+            "self": {
+              "href": "https://api.veracode.com/srcclr/v3/libraries/maven:com.amazonaws:aws-java-sdk-s3:1.11.951:"
+            }
+          }
+        },
+        "vulnerability": {
+          "id": "36376",
+          "title": "Path Traversal",
+          "cve": "2022-31159",
+          "cvss2_vector": "",
+          "cvss3_vector": "AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:N",
+          "cvss2_score": 6.4,
+          "cvss3_score": 6.5,
+          "cwe_id": "CWE-22",
+          "_links": {
+            "self": {
+              "href": "https://api.veracode.com/srcclr/v3/vulnerabilities/36376"
+            }
+          }
+        },
+        "vulnerable_method": false,
+        "_links": {
+          "vulnerability": {
+            "href": "https://api.veracode.com/srcclr/v3/vulnerabilities/36376"
+          },
+          "workspace": {
+            "href": "https://api.veracode.com/srcclr/v3/workspaces/cf0a043f-b752-4a2b-ade8-a36fe3ae0285"
+          },
+          "html": {
+            "href": "https://sca.analysiscenter.veracode.com/teams/dppFqb3/issues/vulnerabilities/12345"
+          },
+          "self": {
+            "href": "https://api.veracode.com/srcclr/v3/issues/8ef6552a-bd7b-4f18-8107-12345"
+          }
+        }
+      }
+    ]
+  }
+}
