--- unittests/scans/veracode_sca/veracode_sca.json
@@ -1,1 +1,0 @@
-{"_embedded":{"issues":[{"id":"ddcc6e1b-3ed9-45c8-b77a-ead759fb5e2c","site_id":129556889,"created_date":"2022-07-29T05:13:00.924+0000","issue_status":"open","issue_type":"vulnerability","ignored":false,"severity":8.8,"workspace_id":"12345","project_id":"12345","project_name":"some-project","project_branch":"master","library":{"id":"maven:org.apache.calcite.avatica:avatica-core:1.11.0:","name":"Apache Calcite Avatica","version":"1.11.0","release_date":"2018-03-06","latest_version":"1.22.0","latest_version_release_date":"2022-07-26","direct":true,"transitive":false,"_links":{"self":{"href":"https://api.veracode.com/srcclr/v3/libraries/maven:org.apache.calcite.avatica:avatica-core:1.11.0:"}}},"vulnerability":{"id":"36527","title":"Arbitrary Code Execution","cve":"2022-36364","cvss2_vector":"(AV:L/AC:L/Au:S/C:P/I:P/A:P)","cvss3_vector":"AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H","cvss2_score":4.3,"cvss3_score":8.8,"cwe_id":"CWE-665","_links":{"self":{"href":"https://api.veracode.com/srcclr/v3/vulnerabilities/36527"}}},"vulnerable_method":false,"_links":{"vulnerability":{"href":"https://api.veracode.com/srcclr/v3/vulnerabilities/36527"},"workspace":{"href":"https://api.veracode.com/srcclr/v3/workspaces/123456"},"html":{"href":"https://sca.analysiscenter.veracode.com/teams/X33hjMQ/issues/vulnerabilities/12345"},"self":{"href":"https://api.veracode.com/srcclr/v3/issues/12355"}}}]},"_links":{"self":{"href":"https://api.veracode.com/srcclr/v3/workspaces/12345/issues?type=vulnerability&project_id=1234&page=0&size=200&sort=id,desc"}},"page":{"size":200,"total_elements":1,"total_pages":1,"number":0}}
